requirements.txt
bottle==0.12.19
bottle-websocket==0.2.9
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
Eel==0.14.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
google-api-core==2.7.2
google-auth==2.6.5
google-cloud-texttospeech==2.11.0
googleapis-common-protos==1.56.0
greenlet==1.1.2
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
proto-plus==1.20.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.8
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
